Move CreditNoteLineItem to its own file
